Move generateId to meta
