--- packages/db/src/helpers/index.ts
@@ -1,4 +1,0 @@
-import { logger } from "@truffle/db/logger";
-const debug = logger("db:helpers");
-
-export { generateId } from "./generateId";
